listView

+++ CelebinoAndroidApp/app/src/main/java/ifpb/edu/br/celebinoandroidapp/Interfaces/ApiInterface.java
@@ -18,10 +18,13 @@ import retrofit2.http.Path;
 public interface ApiInterface {
     @POST("user/login")
     Call<User> login(@Body Login login);
+
     @POST("user/logi")
     Call<User> registration(@Body Login login);
+
     @GET("garden/user/{id}")
-    Call<List<Garden>> getGardens(@Path("id") int id);
+    Call<List<Garden>> getGardens(@Path("id") long id);
+
     @GET("garden/")
     Call<List<Garden>> getAllGardens();
 }
